Warning instead of failing when missing signing properties.
Issuing a warning instead of an exception when signing properties are
not set means that the build won't fail if you build on a system that
doesn't have the signing keys available. This allows for setups where
packages are generated and signed by an automated build system while
still allowing developers to compile the code without the need to set up
a keyring.