Add support of man-page producers.
Lintian wants a man page entry for every executable included into the
package and it wants them gzipped. It's inconvenient to add another
maven plugin to just gzip a man page entry during build. In addition,
it's not so easy to find a correct location for a man page using only
the Debian Policy Manual.

This patch provides a new `man-page` producer type that can guess
correct location for a page and gzip it. Unfortunately, it looks
impossible to gzip a page with maximal compression level as lintian
wants.

In addition, it removes some code duplication by introducing the
`Producers` package-private class to hold common constants and methods
and the `ProducerFactory` class to create producers from run-time
configuration in the ant task and the maven plugin.